Fix handling of single-character FS
Per POSIX spec (and the various implementations), a 1-char FS should be
handled specially as a straight string split and not a regex. However,
I had overlooked this, and was only doing this in the case of "\\".
This fixes things like FS="|" as reported by @shah in #29 (thanks!).

Bumped up minor version, seems like a fairly significant change rather
than just a patch fix.

Fixes #29.