Addresses some IDE warnings relating to GoDoc comments
This addresses some warnings mostly around the convention of the first
word of a godoc being the same as the exported symbol. This also
addresses some typos and some micronits. Any or all of this can be
dismissed if not welcome. I'm just raising this as noticed while
reviewing the code!

Signed-off-by: Adrian Cole <adrian@tetrate.io>