Test performance of I/O-heavy code
Which of course is much more common in AWK...

The program is:

{ for (i = 1; i <= NF; i++) counts[tolower($i)]++ }
END { for (k in counts) print k, counts[k] }

Though only the main actions part is translated to bytecode, not the
END part.

The results are 2.5s for the bytecode version vs 3.1s for original.
Much less impressive than the 3x for non-I/O bound code.

Mawk is still 1.2s (!) so it's doing something right.