Show parse errors in a more standard format
The format is now "filename:line:col: error message" followed by the
source line and caret pointer. I also got rid of the "------" noise.
For example:

$ ./goawk -f s.awk
s.awk:1:12: expected expression instead of }
BEGIN { x* }
           ^

This is closer to how gawk and gcc do it, and is more helpful (includes
the filename) as well as easier to parse for IDEs. Compare to gawk:

$ gawk -f s.awk
gawk: s.awk:1: BEGIN { x* }
gawk: s.awk:1:            ^ syntax error

And gcc:

$ gcc misc/vm-direct.c
misc/vm-direct.c:21:18: error: missing terminating > character
   21 | #include <stdio.h
      |                  ^

It was a little tricky to implement, as ParseProgram doesn't take a
slice of files/filenames, and I didn't want to change the signature
or add a new parse function. So we have to re-read the source files
to count lines till we get to the right filename.

Fixes #43