Support multi-char and regex RS
Operates basically the same as Gawk's RS, except that the ^ and $
anchors match the beginning and end of line, not file. As the Gawk
manual says, "It is thus best to avoid anchor metacharacters in the
value of RS."

Note that I've pulled in more Gawk tests to test this behavior as well.

Fixes #44