Revert using char instead of byte indexes for length/substr/index/match (#95)
The reason is because the new O(N) behavior of these functions was just
too problematic, and caused "accidentally quadratic" issues such as:
https://github.com/benhoyt/goawk/issues/93

A JSON-processing script that previously took 1 second now took over 8
minutes. That's not tenable. The expectation is clearly that these functions
are O(N).

This commit reverts b7ec795b6716aa2159907cce7a54351dc78b0788, but
leaves in the new bytes tests that were added.

Fixes #93