Reduce maxResolveIterations to manageable time; add test for it
Currently if you do actually have thousands of functions with
unresolved types it causes O(N^2) behavior (and takes a looong time),
so for now we significantly reduce maxResolveIterations to something
that only takes a small amount of time (500 takes about 100ms).

Later we can implement a proper topological or similar sorting
algorithm so these are processed in the correct order, avoiding the
O(N^2) worst case.