CSV support (#107)
This implements proper CSV support as described in
https://github.com/benhoyt/goawk/issues/105

Documentation for the CSV feature is included in the csv.md file.

All the features there are implemented. Note that assignment using
the new named field is not supported, for example you can't do
`@"name" = "xyz"` to update a field; I may add this later, though it
doesn't seem particularly useful to me.

Fixes #105.