flushAll() stops flushing on first error (#209)
The && operator causes conditional evaluation (aka "short-circuiting").
This means if the left-hand-side is false the right hand side is
not evaluated. If any call to interp.flushWriter returns false then
all streams following that stream in the range, as well as
interp.output, will not be flushed.

Removes the type cast to the flusher interface because that is done
in interp.flushWriter. This check would be performed twice for any
interp.output value whose type implements the interface.

Fixes #208