Parse interpTest flags and implement "!windows" flag
The !windows flag will cause both the external awk tests and the goawk tests
to be skipped if the tests are running on Windows.

Signal-based test cases were failing on windows for goawk because Windows
does not support POSIX signals. These are simulated on cygwin which is
what MSYS2 appears to be using. MSYS2 is preinstalled on the Windows image
which the CI runs.

My best guess is that the MSYS2 kill command is forcing the child process to
terminate with (signal << 8) as the exit status.

Additionally, the gawk in the Windows CI image is using cmd.exe and not sh.
Naturally, there are no exec or kill commands available from cmd.exe.

Ironically, the signal exit status behaves predictibly for gawk on cygwin.

Stricter flag parsing exposed some interpTest entries that were missing
a comment char ('#').