Fix rare race in pipe-to-command close() result test (#213)
The bufio.Writer does not write to the Cmd pipe until Flush is
called when the Cmd is Wait-ed on. This causes a Flush error but
not a Wait error. This only happens if the Cmd starts and exits
before the Flush can be called by goawk.

Fix this by forcing the spawned "sh" command to read a line of
input before exiting or sending a signal to itself.

Automated tests ran into another problem:

The windows GitHub runner installs gawk using Chocolatey and this
package of gawk does not use "sh" to execute commands, instead it
uses "cmd.exe". Because "read" is a "sh" builtin command this will
not work for windows tests of gawk and probably other awks.