Use C.UTF-8 as locale when invoking external awk program
The en_US.UTF-8 locale is not guaranteed to be installed, and it quite
often isn't by default.

The C.UTF-8 locale was introduced as a Debian-specific locale, but was
later upstreamed into glibc 2.35 [1].

This patch fixes failures when running interpreter tests on systems
where the en_US.UTF-8 locale is not present. It also changes the locale
on goawk_test.go, similar to !174 [2].

MacOS does not support the C.UTF-8 locale, so we keep using en_US.UTF-8
in that case.

1: https://sourceware.org/bugzilla/show_bug.cgi?id=17318
2: https://github.com/benhoyt/goawk/pull/174