Standardise float64 -> uint64 conversions on macOS/ARM
See https://github.com/golang/go/issues/52943

Go Spec says "In all non-constant conversions involving floating-point
or complex values, if the result type cannot represent the value the
conversion succeeds but the result value is implementation-dependent."

On x64 this does uint64(int64(f)), but on ARM it yields 0. The former
behaviour seems more intuitive, so run with that.