Use WaitDelay to ensure system() exits quickly when context is cancelled (#270)
I don't think this is an exhaustive fix, but it fixes the main issue of
system() never returning when the context is cancelled (due to
stdout/stderr pipes being passed to the child process).

This PR also bumps up the minimum Go version required in go.mod from
1.18 to 1.20, as `Cmd.WaitDelay` was only introduced in Go 1.20. That's
fairly old by now, so this seems reasonable.

Fixes #122